Refactor recipe upload to use a server function
