+++ src/pages/recipes/recipeUpload.constants.ts
@@ -1,0 +1,2 @@
+export const RECIPE_UPLOAD_MAX_PHOTO_COUNT = 8;
+export const RECIPE_UPLOAD_MAX_PHOTO_BYTES = 10 * 1024 * 1024;
